68-install-bootstrap-sass
I installed and prepared `bootstrap-sass` gem for future usage. Also replaced application.css with application.css.scss and clean it up.

If we decide to use Sass we need to giveup of using sprockets with css files because none of other stylesheets will be able to acccess Bootstrap mixins and variables.

Also added required dependency line in javascript manifest.